+++ tests/test_text_to_audio_models.py
@@ -14,8 +14,14 @@ from llmmaster.text_to_audio_models import ElevenLabsTextToSpeech
 from llmmaster.text_to_audio_models import VoicevoxTextToSpeech
 from llmmaster import LLMMaster
 
+
+API_KEY = '''
+'''
+
+
 TEST_OUTPUT_PATH = 'test-outputs'
-PROMPT = "Do not concentrate on the finger, or you will miss all that heavenly glory."
+PROMPT_EN = "Do not concentrate on the finger, or you will miss all that heavenly glory."
+PROMPT_JA = "こんにちは。今日もいい天気ですね！"
 
 
 @pytest.fixture
@@ -30,9 +36,10 @@ def test_openai_text_to_speech_basic(run_api):
     test_cases = []
 
     for voice_pattern in OPENAI_TTS_VOICE_OPTIONS:
-        test_case = {'provider': 'openai_tts', 'prompt': PROMPT, 'voice': voice_pattern}
+        test_case = {'provider': 'openai_tts', 'prompt': PROMPT_EN, 'voice': voice_pattern}
         test_cases.append({'name': f'openai_tts_{voice_pattern}', 'params': test_case})
 
+    master.set_api_keys(API_KEY)
     for case in test_cases:
         master.summon({case['name']: master.pack_parameters(**case['params'])})
 
@@ -74,13 +81,14 @@ def test_openai_text_to_speech_various_options(run_api):
     master = LLMMaster()
 
     test_cases = [
-        {'name': 'openai_tts_v1_echo_voice_slow', 'params': {'provider': 'openai_tts', 'prompt': PROMPT, 'voice': 'echo', 'speed': 0.5}},
-        {'name': 'openai_tts_v2_fable_voice_fast', 'params': {'provider': 'openai_tts', 'prompt': PROMPT, 'voice': 'fable', 'speed': 1.5, 'response_format': 'opus'}},
-        {'name': 'openai_tts_v3_onyx_voice_model_tts_1', 'params': {'provider': 'openai_tts', 'prompt': PROMPT, 'voice': 'onyx', 'model': 'tts-1', 'response_format': 'aac'}},
-        {'name': 'openai_tts_v4_nova_voice_model_tts_1_hd', 'params': {'provider': 'openai_tts', 'prompt': PROMPT, 'voice': 'nova', 'model': 'tts-1-hd'}},
-        {'name': 'openai_tts_v5_shimmer_voice_response_format', 'params': {'provider': 'openai_tts', 'prompt': PROMPT, 'voice': 'shimmer', 'response_format': 'opus'}},
+        {'name': 'openai_tts_v1_echo_voice_slow', 'params': {'provider': 'openai_tts', 'prompt': PROMPT_EN, 'voice': 'echo', 'speed': 0.5}},
+        {'name': 'openai_tts_v2_fable_voice_fast', 'params': {'provider': 'openai_tts', 'prompt': PROMPT_EN, 'voice': 'fable', 'speed': 1.5, 'response_format': 'opus'}},
+        {'name': 'openai_tts_v3_onyx_voice_model_tts_1', 'params': {'provider': 'openai_tts', 'prompt': PROMPT_EN, 'voice': 'onyx', 'model': 'tts-1', 'response_format': 'aac'}},
+        {'name': 'openai_tts_v4_nova_voice_model_tts_1_hd', 'params': {'provider': 'openai_tts', 'prompt': PROMPT_EN, 'voice': 'nova', 'model': 'tts-1-hd'}},
+        {'name': 'openai_tts_v5_shimmer_voice_response_format', 'params': {'provider': 'openai_tts', 'prompt': PROMPT_EN, 'voice': 'shimmer', 'response_format': 'opus'}},
     ]
 
+    master.set_api_keys(API_KEY)
     for case in test_cases:
         master.summon({case['name']: master.pack_parameters(**case['params'])})
 
@@ -122,8 +130,9 @@ def test_elevenlabs_text_to_speech(run_api):
     judgment = True
     master = LLMMaster()
 
-    test_case = master.pack_parameters(provider='elevenlabs_tts', prompt=PROMPT)
+    test_case = master.pack_parameters(provider='elevenlabs_tts', prompt=PROMPT_JA)
 
+    master.set_api_keys(API_KEY)
     master.summon({'elevenlabs_tts': test_case})
 
     print(f'elevenlabs_tts = {master.instances["elevenlabs_tts"]}, {master.instances["elevenlabs_tts"].parameters}')
@@ -158,10 +167,8 @@ def test_voicevox_text_to_speech(run_api):
     judgment = True
     master = LLMMaster()
 
-    prompt = 'こんにちは。今日もいい天気ですね！'
-
-    case1 = master.pack_parameters(provider='voicevox_tts', prompt=prompt, speaker=1)
-    case2 = master.pack_parameters(provider='voicevox_tts', prompt=prompt, speaker=2)
+    case1 = master.pack_parameters(provider='voicevox_tts', prompt=PROMPT_JA, speaker=1)
+    case2 = master.pack_parameters(provider='voicevox_tts', prompt=PROMPT_JA, speaker=2)
     master.summon({'case1': case1, 'case2': case2})
 
     for name, instance in master.instances.items():